Add files via upload
This is my first task at Jensen Yrkeshögskola (collage), it was at the beginning of fall -2021. Our task was to build a responsive website, with pictures, dark theme switch/button, bootstrap, grid system, CSS, JavaScript, drop-down menu, navbar, design (like the Facebook and Twitter logos in the footer, colors, borders, rounded corners, etc). We also did a ''contact'' page. 

Pretty basic website but enough to look clean and simple for a first project. The goal was not to make the most beautiful website but to learn basic functions, design, etc. 

All in all, as the first project with no prior experience of coding/programming, I'm quite happy with the result.